move to single client recipe with no master
recipe will now do the following:

- if local authkey file exists, move on
- if not, search for other nodes that have it in their node hash
- if none, generate it yourself and set it to your node hash
- if at least one is found, grab the authkey fromit's node hash and then
  set it to your own. decode, and drop it into local file

This way, we only need to have a single recipe with no master/client.
All clients become potential search targets to provide the key to new
clients.

NOTE: there is an assumption that you will run this on a single node
first so that it can set the authkey, with no race condition.