Simplify roles a bit
We now have an attribute that helps us know when a node is a founder or
not. The end goal will be to get rid of the founder role and have the
wrapper (for instance, crowbar) set this attribute on one node.

This also allows installing pacemaker before corosync, avoiding issues
where corosync doesn't know about pacemaker.