Add helpers to be able to "synchronize" nodes in the cluster
When non-founder nodes call:
  PacemakerHelper.wait_for_mark_from_founder(node, cookbook_name, "pki-sync", node[:keystone]["crowbar-revision"])

they will be blocked until the founder calls:
  PacemakerHelper.set_mark_if_founder(node, cookbook_name, "pki-sync", node[:keystone]["crowbar-revision"])

This also comes with a LWRP, to enable synchronization in the second
phase of a chef run. It can be used to allow the founder to execute some
resources first. For instance:

  - before the critical code that the founder should run first:

    crowbar_pacemaker_sync_mark "wait-for-endpoints-mark" do
      mark "endpoints"
      revision node[:keystone]["crowbar-revision"]
      action :wait
    end

  - after the critical code that the founder should run first:

    crowbar_pacemaker_sync_mark "set-endpoints-mark" do
      mark "endpoints"
      revision node[:keystone]["crowbar-revision"]
      action :create
    end

If the revision parameter is not passed, then
node[cookbook_name]["crowbar-revision"] will be used, which is
convenient for crowbar integration.

The name of the crowbar_pacemaker_sync_mark can also be used to
automatically generate the mark parameter and make it optional. For
instance:

    crowbar_pacemaker_sync_mark "wait-test-def" do
      action :wait
    end

    crowbar_pacemaker_sync_mark "create-test-def" do
      action :create
    end