Add a crowbar-pacemaker wrapper cookbook and switch to it
This cookbooks wraps all the use of upstream cookbooks, in order to
allow proper crowbar integration, as well as integration between two
independent cookbooks.

Concretely, we provide an easy way to install haproxy on pacemaker
clusters.

Note that we always install haproxy on cluster nodes.If we were not doing this,
then some nodes would not have it installed when the resource is first declared
and these nodes will have this error:
    warning: services_os_action_execute: Cannot execute '/etc/init.d/haproxy': No such file or directory (2)
The error is fixable with "crm resource refresh", but we might as well just
install the package.