Only load crm-initial.conf when the file got changed
There is no point in loading it if it's the same as before, since we
already imported these settings.

We load it immediately because we know that corosync is running already,
and it's a critical file that we do not want to fail loading because of
a later crash in chef-client.