Allow specifying an arbitrary public name for the public virtual IP
This is important because end users will see the public FQDN
(cluster-foo.domain, with foo being the name of the pacemaker proposal),
which can be considered ugly but also an implementation detail. This
happens for endpoints, for instance.

The public FQDN also matters for SSL certificates, since the SSL
certificates need to be valid for that hostname.

So if we want to hide this public FQDN, having the cloud operator define
a CNAME in the DNS is not enough, we also need to have the barclamp know
about the public name picked by the cloud operator. Hence this new
setting.

This is obviously HAProxy-specific.