Adapt attributes created by crowbar for per-node STONITH
For per-node STONITH, we have a hash where the keys are the FQDNs of
each cluster member. It turns out that when configuring the resources in
pacemaker, what we are using is actually the hostnames of each cluster
member (ie, FQDN without the domain name).

So we translate what crowbar gives us to what pacemaker expects.