corosync: Ensure that we always use authkey from founder when in Crowbar
When the corosync cookbook is running orchestrated by Crowbar, then we
know that there will be a member of the cluster that is the pacemaker
founder, and we will always use the authkey from that node.

When not running orchestrated by Crowbar, then we will just use the
first authkey we find.