correctly parse empty parameter values in CIB objects
For a CIB object definition such as:

    primitive database-config-default-fs ocf:heartbeat:Filesystem \
            params device="/dev/drbd/postgresql" directory="/var/lib/pgsql" fstype="" \
            op monitor interval="10s"

when #extract_hash attempts to extract the 'params' section,
Shellwords.split will return an Array whose last element is

    fstype=

not

    fstype=""

as you might expect.  Therefore we need to allow for an empty string to
the right of the equals sign, otherwise parsing fails.