don't ignore synchronization failures by default
The "fatal" attribute of crowbar_pacemaker's sync_mark LWRP defaulted to
false, so synchronization failures effectively got ignored - the only
difference being an extra 60-second delay and a warning in the logs.

One symptom of ignoring synchronization failures is that Pacemaker
resources get created and started before setup is complete on all nodes.
This in turn causes fencing of nodes with incomplete setup when
Pacemaker attempts to start up the resources on those nodes.

It's more user-friendly if rather than fencing, the chef-client run
simply fails, since this gets bubbled up to the UI without unnecessary
failovers or interruptions to existing services running on the node in
question.

Of course, any synchronization which is advisory-only can still specify

  fatal false

in the resource block.