Use more descriptive msg on sbd init check failure
Stonith + SBD deployment can fail due to SBD not being created on the
shared disk. The message does not really tell what went wrong and what
should be done to fix it. Use a more descriptive one.

The SBD creation is a destructive op and a typo could wipe the user
data, so this is the reason that sbd create is not run by default.

An alternative would be to run the sdb create tonly after user types
the volume name into the input field in the UI twice, for confirmation.