Fix conversion of per node stonith parameters
This converts the per node stonith parameters from Chef::Node::Attribute to a
"normal" ruby hash when stripping the domain from the node names. It seems
assigning a hash of Chef::Node::Attribute objects to the node attributes
doesn't work. When accessing the values later on (in pacemaker::stonith) they
just evaluated to nil.

https://bugzilla.novell.com/show_bug.cgi?id=887244