Raise header bufsize to 32kb (bnc#900950)
It seems with multiple regions, the 24kb header size for
PKI tokens are exhausted pretty quickly. Raise to 32kb.
maxrewrite would be raised by that by default to 16kb, but
we don't need such a high value (since we don't rewrite
any of the large headers). So keep it at 8kb.