immediately abort chef-client if maintenance mode is set
The reasoning is explained in the comments.  I'm not sure why I ever
thought it was a good idea to allow a chef-client run to proceed if
maintenance mode is set.  It's not even as if that approach could ever
restore an ill node to full health, because we were deliberately
checking whether the maintenance mode was set prior to the chef-client
run, and if so, leaving it set.  This way, if a node is left in
maintenance mode, it will be discovered sooner, resulting in the cloud
operator being alerted to a degraded cluster sooner.