remove voodoo used for instantiating a new Chef resource
We already have @new_resource.class, so use that instead of doing
surgery on class names.  This is cleaner, and anyway the surgery breaks
with newer versions of Chef where the provider class is a singleton
class with a human-readable string as the class name, rather than
Chef::Provider::Foo.