crowbar-pacemaker: Do not create remote resources until it's possible
It could happen that the chef search returns nodes which have not been
fully setup yet. For instance, if chef crashes while running.

When this happens, because the remote node is not reachable through
pacemaker_remote, pacemaker will try to fence it, which will just fail
due to the lack of STONITH at this point in time.