Get rid of pacemaker-remote-delegator role
This is slightly confusing for the user, and it's important that it's
filled correctly (with founder), which makes it a bit hard to use.

Instead, just re-run pacemaker-cluster-member after running
pacemaker-remote. This will just work fine anyway because the first time
we pass in pacemaker-remote, the remote nodes won't have set the
attributes that declares they're ready for remote, and so only in the
second pass of pacemaker-cluster-member will they be added as remote
nodes.

Also, when there's no remote node, just hack the element order to avoid
runnning twice uselessly.