crowbar-pacemaker: Add location constraint for haproxy & apache
Because the pacemaker resources for haproxy and apache are created here
in crowbar-pacemaker, we, in theory, shouldn't assume there that these
two services will be used for OpenStack and so we can't always just add
the location constraint to tell pacemaker to only run these on
controller nodes.

However, we do need the constraints when it's OpenStack. So crowbar will
add a flag to tell us to add these constraints, and when it's there, we
do what we're asked to do.