pacemaker: Compare cib objects definitions as one-line definitions
This matters because what "crm configure show" gives us and what the
cookbooks ask for can, in theory, be multi-lines, but not necessarily
with the same amount of lines.

A common example is:

location l-cl-g-nova-compute-compute cl-g-nova-compute resource-discovery=exclusive \
        rule 0: OpenStack-role eq compute

crm configure show will be on two lines, but the cookbooks have a one-line
definition. But they're really the same definition, so no need to change anything.

An alternative approach would be to store multi-line definitions as
one-line definitions; unfortunately, some code (like the parsing of
op/meta/params of resources) assumes and depends on multi-line output
from "crm configure show".