pacemaker: Copy what crm does by adding interval=0 for primitive ops
It seems recent version of crm (or maybe it was always like this?)
always add interval=0 to ops of primitives that have attributes except
for an interval.

This unfortunately results in creating some diff between the desired
definition and the existing definition. We cannot really remove
interval=0 when parsing an existing definition (since it could actually
be part of the desired definition), so just copy crm's behavior when
creating the op string.