pacemaker: Only call "crm configure show" once per transaction
Previous code was using ::Pacemaker::CIBObject.exists? for each object
in the transaction, which calls "crm configure show" that many times.

crm being a python tool, it can take some time to start, and doing that
so many times is not very optimal. So just run it once and filter the
output to get a list of existing objects.