crowbar-pacemaker: Add restart_crm_resource flag to service provider
This flag tells us to use crm_resource for restarting the service. This
is useful for resources that are defined with an OCF agent: in that
case, we can't proxy the restart action to a LSB init script or systemd.

The usual downside of going with the crm/pacemaker tools for this is
that the resource could move to another node, or could trigger changes
in the cluster due to constraints. To avoid this, we use "crm_resource
--force-stop / --force-start" which actually bypasses the cluster.