crowbar-pacemaker: Fix fencing of remote nodes on creation of primitive
We were creating the primitive with [:create, :start]. But the devil is
in the details: with :create, we force the primitive to be created with
target-role="stopped", which results in pacemaker ensuring that the
resource is stopped. That means fencing.

So instead, we put the primitive in a transaction, which leads to the
primitive to be started automatically.