Don't try to blank remote_host DRBD attribute
If a node gets fenced, for some reason the pacemaker-cluster-member
role can end up being removed, in which case the Chef search which
is used to find the node's DRBD peer will not find it, and the recipe
will try to set its remote_host attribute to nil, causing an exception.
Instead we now issue a warning and keep the previous value.

However if we never got a non-nil value in the first place, still raise
an exception because something bad must have happened.