crowbar-pacemaker: Add sync marks to apache HA setup
The mod_status configuration happening at the beginning of this recipe
needs to be run on all nodes before setting up the apache HA resources.

This is quite important when upgrading from tex, as we might have an old
version of apache before the upgrade with a different syntax in the
mod_status config file. If we don't have sync marks there, then the
resource might start on nodes where the mod_status config file was not
updated, resulting in apache just failing to start due to invalid
syntax.

(And in general, it's the right thing to do anyway)

It's worth mentioning that the revision we're picking here is "good
enough", but not fully correct: technically, we don't know which
barclamp is using apache so we don't know which revision to use; but the
first time we set apache up, this sync mark will be empty so the sync
mark magic will happen anyway.

(cherry picked from commit ea764b2e2e247ae9698cafe50a866e883d1dc4b4)