Added http-keep-alive option to haproxy defaults
We were using the http-request for haproxy with a 5s timeout. This
option only addresses the initial part of the HTTP negotiation. Once
the HTTP session is established we should use a longer timeout
for idling connections with the http-keep-alive option. This option
if not set gets the same value as http-request timeout.

The defaults for http-request and http-keep-alive have been increased
to 10 seconds. This seems to be in sync with OpenStack documentation
as well as other vendors implementations.

This patch will help fix problems with 'nova boot --poll', because
the connections usually ended at the same time a new request was
coming, leading to connection refused errors.

References:
- https://trello.com/c/m9a3MQYS
- https://cbonte.github.io/haproxy-dconv/configuration-1.5.html#4.2-timeout%20http-keep-alive