corosync: Rework "block autostart" to not rely on chef-client on boot
Right now, the feature to block autostart of corosync was relying on
chef to start corosync (or pacemaker, on newer platforms) on boot.

In order to move away from the requirement of having chef-client running
on boot, this is reworked so that this relies on a new service (for
sysvinit) or a overridden corosync service (on systemd) for the whole
logic. This way, the feature work just fine without any chef run on
boot.

For SLE12, it also means we can always enable pacemaker now, since it
will fail to start due to corosync if needed.