haproxy: Make the sections for haproxy config persistent on each node
Until now, we were saving the sections as a default attribute, so that
they would be removed on every new run, resulting in the config being
regenerated from scratch. The downside of this is that all active roles
that use a loadbalancer resource must always be used in the chef run,
otherwise we lose part of the config. And this goes against our goal of
running only the strict minimal recipes when required.

Instead, we should just add a :delete action and use it.