Do not use node.ipaddress, but check for 'admin' range IP
node.ipaddress is set to the address of the interface that is also used for default router.
I.e. when the node has an address in the public network, node.ipaddress will be set to the address assigned to the public interface.
As we don't want drbd to use the public network for replication, use the barclamp_library to get the IPs on the admin network.