crowbar-pacemaker: Split setup from definition of the DRBD resources
When we apply the pacemaker proposal, we want to setup everything for
DRBD, if only because we need to claim a disk (so it's not something we
want to postpone). However, we don't know yet the resources that will be
defined.

When we apply later on a proposal that uses a DRBD resource, then the
cookbook from that proposal should include the small recipe to configure
the DRBD resource, but doesn't have to do the setup of DRBD itself.

The benefit is that we can skip running the pacemaker role when applying
database, as database will include the required recipe for DRBD.