crowbar-pacemaker: Translate crowbar-specific stonith in rails app
We were doing this in the crowbar-pacemaker cookbook before, but we can
already do it in the rails app for free, and it has the benefit that the
configuration is persistent in chef runs, even if the crowbar-pacemaker
recipes are not used. This matters for nova compute HA since we check
some of these attributes there, and we so far relied on the
crowbar-pacemaker recipes being run.

This also removes some validation from the cookbook, but this was also
done in the rails app.