drbd: Do not use drbd-overview while creating the drbd resource
Due to fbd12dfd, the drbd daemon may not run (as we let pacemaker start
it, and for the first resource, it will therefore not run yet), so
drbd-overview will not work.

We were using drbd-overview to make sure that all steps are working;
without it, we will just fail a bit later.

If there's a need to wait for drbd to be ready, there's a new :wait
action for the resource.