crowbar-pacemaker: Remember crowbar-specific STONITH mode too
Since we translate the STONITH mode in the rails app, the mode attribute
is now always one of sbd/per_node/shared. However, in the
crowbar-pacemaker cookbook, we still need access to the untranslated
mode to do some additional steps in the setup (especially for libvirt,
which is translated to per_node).