crowbar-pacemaker: Rework how we setup mutual ssh keys
First, we always use the default attributes for the provisioner access
keys as a basis, instead of using normal attributes. This ensures that
we always have the keys from the provisioner proposal and the keys from
the cluster nodes and nothing else. Before, we could have keys from
removed cluster nodes still left around.

Also, we only save the node if there's really a change in the attribute.