crowbar: Save founder name in the proposal role
This replaces the boolean attribute that we were setting on each node.
On top of the simplification of the code, this allows us to not search
for the founder node since we can simply load it.

It will also enable us to move the sync marks to not be
chef-attribute-based, but cib-attribute-based without a search or a
load, since we will only need the name of the founder in that case.

(cherry picked from commit 5eb1716377301d96cdd79c621b1f4dad3f75d424)