Add a helper for pacemaker_clone meta attributes (bsc#965886)
Return a commonly appropriate hash for the meta attribute of a
pacemaker_clone resource.  This will allow us to remove a lot of
duplication between identical meta attributes across many
pacemaker_clone resources.

We were already setting clone-max in many places to limit the number
of clones to the number of corosync nodes, otherwise the UI would get
noisy by showing inactive clone instances on remote nodes which the
clones are constrained from running on.  Conversely we were also
restricting the number of compute-only clones to the number of remote
nodes, avoiding inactive clones showing on the corosync nodes.

In addition to that common pattern, we also want to interleave our
clones:

By default, Pacemaker clones aren't interleaved.  This means that if
Pacemaker wants to restart a dead clone instance, and there is an
order constraint on that clone, it will do the same restart on all
other nodes, even if all the others are healthy.

More details on interleaving are here:

    https://www.hastexo.com/resources/hints-and-kinks/interleaving-pacemaker-clones/

This behaviour is far more disruptive than we want.  For example, in

    https://bugzilla.suse.com/show_bug.cgi?id=965886

we saw that when a network node dies and Pacemaker wants to stop the
instance of cl-g-neutron-agents on that node, it also stops and
restarts the same clone instances on the healthy nodes.  This means
there is a small window in which there are no neutron agents running
anywhere.  If neutron-ha-tool attempts a router migration during this
window, it will fail, at which point things start to go badly wrong.

In general, the cloned (i.e. active/active) services on our controller
and compute nodes should all behave like independent vertical stacks,
so that a failure on one node should not cause ripple effects on other
nodes.  So we interleave all our clones.

(There are corresponding commits to the other crowbar repos for other
clones.)