crowbar-pacemaker: Replace sets with one resource in order_only_existing
In "only existing" order constraints, we want to replace sets with only
one resource like:

  ( foo ) bar

with:

  foo bar

The rationale is that crm does this, and when we compare the resource
definition we want to have with the existing resources in crm, we need
to make sure to have the same definition syntax.

Of course, this doesn't have any other impact because the behavior for
pacemaker is the same.