drbd: Set node-id in resource config
Some of the failing mkcloud HA tests (e.g. [1])) have the following chef
failure:

  Failure: (119) No valid meta-data signature found.

accompanied by an alert in the logs:

  ambiguous node id: meta-data: 0, config: 1

This patch sets the node IDs explicitly in the resource config file as
recommended in this mailing list post[2] and documented for drbd
9.0[3] to attempt to prevent these meta-data errors. It also adds a
connection stanza as that may also be affecting drbd's ability to
generate metadata[4]. Protocol C indicates "Writes to the DRBD device
complete as soon as they have reached the local and all remote disks"
and is the protocol used in the example configuration file in the
documentation. We could consider making this configurable later.

[1] https://ci.suse.de/job/openstack-mkcloud/56301/consoleText
[2] https://docs.linbit.com/doc/users-guide-90/re-drbdconf/
[3] http://lists.linbit.com/pipermail/drbd-dev/2017-January/003822.html
[4] http://lists.linbit.com/pipermail/drbd-dev/2017-January/003823.html