haproxy: Add sync marks
Without this patch, the keystone proposal often causes a crm resource
failure when it tries to start the haproxy service under pacemaker:

  Failed Actions:
  * haproxy_start_0 on d52-54-77-77-01-02 'not installed' (5): call=64, status=Not installed, exitreason='none',
      last-rc-change='Fri Apr 28 09:53:22 2017', queued=0ms, exec=92ms

When we examine the zypper history of both controller nodes, we can see
the haproxy package may be installed by 09:53:22 on one of the nodes,
but is only installed a few seconds later on the other node, leading
pacemaker to report a failure. This patch adds sync marks (modeled after
the apache recipe in the same cookbook) to ensure pacemaker does not try
to start haproxy until both nodes are ready.