make synchronization failure messages more helpful
Chef sync_mark synchronization failures are one of the most frequent
source of problems and confusion in SOC.  We constantly receive
support requests where the reported error is of a node timing out
whilst waiting for another node, because users don't realise that that
is not the underlying cause of the error so they don't check the
correct log file.

So tweak the error messages to make it more obvious what's going on,
and to encourage users to check the correct log files to find out and
report the real problem, not just a symptom of it.

We choose to extend the log messages with reference to specific
non-FQDN hostnames, as the FQDN versions would make the messages
unnecessarily long.  If it's type A synchronization (i.e. "wait for
founder"), tell the user which node is the founder.  If it's type
B ("wait for all other nodes") then tell the user which nodes we're
still waiting for.  A nice side-effect of this last change is that we
calculate the remaining nodes via set subtraction rather than by
length comparison, which should be more accurate.