Add a support to use 'backup' option in server line.
'backup' means that given server is only used in load
balancing when all other non-backup servers are unavailable.

'stick on dst' should make sure that destination server is not
left even if it is backup one and non-backup is healthy again.
This should prevent additional disruption once the failing node
is back and running.

We might need a setup (e.g. with mariadb cluster) where we want
to prevent existence of multiple servers being accessed at the same time.