crowbar-pacemaker: Allow overriding start/stop commands for crm_resource
In order to successfully restart a mariadb instance that is part of a
galera cluster we need to cycle it through the Slave/Master roles once.
I.e. instead of doing a --force-stop/--force-start cycle we need to call
crm_resource with --force-demote/--force-promote. This patch enhances
the pacemaker service provider to allow specifying which operations to
call for stopping/starting the resource. (It will default to the old
--force-stop/--force-start behaviour)