pacemaker: Add option to stop managing stateless active/active services
Pacemaker is not strictly required for stateless active/active services,
since orchestration, fencing, and so on is not necessarily useful in
that case.

Disabling this option does simplify the setup in the OpenStack case quite
significantly, but comes with a cost, like pacemaker not restarting
services that have failed.

The default is to keep the same behavior as in the past. This may change
when we implement Restart= for systemd services and if it found that
this approach has no negative impact.