pacemaker: Add missing  operations to the parser
The parser for the "op" part of a resource definition was missing the
"promote" and "demote" operations. This was never noticed up to now as
we didn't use them until we introduced galera support in the database
barclamp. This bug caused calls to "crm config" in every chef-client run
for resources that contain custom settings for the "promote" or "demote"
operation.