crowbar-pacemaker: Reimplement sync marks with pacemaker attributes
The main benefit here is that we don't depend on the chef server
(through searches), which should decrease the load on the admin server
but also make things faster.

There's a slight difference in behavior for the sync marks: only the
founder nodes is waiting for all other nodes on sync; the non-founder
nodes simply wait for the founder because the founder will only set the
mark once all other nodes have already set their mark.

The initial sync mark (about the founder node being setup) is replaced
by a simple attribute on the node, in a similar way that we mark remote
nodes having been setup.

It's worth mentioning that we don't use revisions but only booleans for
the sync marks: revisions will be deprecated in the next commit.

Finally, we convert pre-existing sync marks to this model, which should
make this change non-disruptive.

Co-Authored-By: Vincent Untz <vuntz@suse.com>
(cherry picked from commit e47039c583222c15587bf38234f6b85ba4e368d5)