crowbar-pacemaker: Deprecate usage of revisions in sync marks
Thinking about it, I think it was a mistake to do this: once a sync mark
has been reached once, there's no point in waiting on it again later on
only because the revision is increased: the code is the same, or if the
proposal change triggers some new code that needs a sync mark, then it
should be a new sync mark anyway.

That's also why the pacemaker attribute backend only uses boolean for
sync marks.

Co-Authored-By: Vincent Untz <vuntz@suse.com>