Add support for multiple Corosync rings
Customer has requested the ability to configure an additional corosync
ring to "fully leverage the second bond in their dual bond environment".
The pacemaker cluster nodes will have visibility of one another over
both bonds and, when a failure occurs on the first bond, a cluster
failure will be averted due to the existence of the second ring over
the second bond.

In order to accomplish this, the flat list of single ring attributes
and web UI form fields is replaced by an array and n number of web UI
form fields to support n number of rings. The UI and validation code
restrict the maximum number of rings to 2. If there is ever a need for
for more rings, simply search for '2', 'two', and 'second' and make
the appropriate changes.

* Changes the default transport from UDP to UDPU (the recommended protocol)
* Replaces the flat variables of a single ring with an array of ring
hashes to support n number of rings, currently restricted to 2 by
UI and validation
* Provides a schema migration script
* Fixes UDP by including node_list in conf file regardless of transport
* Adds multiple ring support to crowbar UI, currently restricted to 2
by javascript (if raw is used, validation will occur in barclamp)
* Allows user to select a network other than admin for the first corosync
ring
* Allows user to specify multicast address (previously only in raw mode)
* Adds validation to ensure the members lists are not empty
* Adds validation to ensure all members lists have the same length
* Adds validation to ensure a network is specified
* Adds validation to check if a specified network exists
* Adds validation to check if a specified network is unique
* Adds validation to ensure a multicast address is specified
* Adds validation to check if a specified multicast address is in use
and, if so, suggests one that is available
* Allocates and brings up IP addresses as necessary to ensure that
each node is a member of the specified network
* Updates the firewall port list for all rings
* Opens multicast port - 1 if transport is UDP

(cherry picked from commit 941deec)