crowbar-pacemaker: Reset sync-marks for all nodes
Up to now we only reset the synchronisation marks for the founder node
when re-applying a barlcamp. The reset needs to happen for all nodes of
a cluster. Otherwise we will run into synchronisation issues if e.g. the
first deployment of a barclamp fails and all nodes except of the founder
still have the sync marks set.