pacemaker: Extent extract_hash to deal with multiple entries
It turns out that we can have multiple entries for the same data type.
For instance, when using ms, we need to define multiple "op monitor"
with different roles:
  https://clusterlabs.org/doc/en-US/Pacemaker/1.1/html/Pacemaker_Explained/_monitoring_multi_state_resources.html

To adapt to that, when we detect this case, we return an array of
hashes, instead of just a hash. Note that this is, to my knowledge, only
possible for op, so we raise an exception when this happens for
something else.

In all other cases, the behavior is the same. That is: we return a
single hash (which is empty when there's no match).