pacemaker: Adapt op_string to deal with multiple entries of same type
It turns out that we can have multiple entries for the same data type.
For instance, when using ms, we need to define multiple "op monitor"
with different roles:
  https://clusterlabs.org/doc/en-US/Pacemaker/1.1/html/Pacemaker_Explained/_monitoring_multi_state_resources.html

This changes op_string to be able to deal with an array of hashes when
this is the case.

(cherry picked from commit 07e7c4ef42718044044cc4492b60254cb9804c62)