pacemaker: primitive to read global op defaults
The pacemaker_primitive object has been extended in order to add support
for global 'op monitor on-fail' defaults. In this commit we consider the
barclamp schema already supports an option to configure the defaults for
the 'op monitor on-fail' action.

The primitive will read that value from the founder node and set it to
all 'op monitor on-fail' actions. Should the user had configured
manually the defaults (editing the barclamp), then those manually
configured values will prevail.

This code required the 'op monitor on-fail' attribute to be configured
as a Chef::Node::Attribute class. The main reason is that this class
keeps the original values configured in the barclamp as well as the new
ones, so it is possible to roll-back changes.

Some barclamps set this attribute as a Hash object. In that case there's
no possible rollback. The behavior in this case will be to not modify
it, in order not to break anything. The user can still manually set the
values in that particular barclamp to suit their needs, but not as a
global option.